add push.py
