+++ docs_converter.py
@@ -23,41 +23,13 @@ import re  # 정규식 사용
 
 sys.path.append(os.path.abspath('/data/Database Project'))
 import output_DB
+import push
 
 router = APIRouter()
 
 class ConverterPayload(BaseModel):
     doc_type: int
     doc_s_no: int
-
-def push_to_nextjs(file_path, file_name):
-    try:
-        try:
-            logging.info(f"Sending file {file_name} to Next.js using Raw Binary")
-
-            with open(file_path, "rb") as file:
-                response = requests.post(
-                    "http://192.168.50.84:90/api/file_receive",
-                    data=file,
-                    headers={
-                        "Content-Type": "application/octet-stream",
-                        "file-name": quote(file_name)
-                    }
-                )
-            if response.status_code != 200:
-                logging.error(f"Frontend server response error: {response.status_code} - {response.text}")
-                raise HTTPException(status_code=500, detail="Failed to send file to frontend")
-
-            logging.info(f"File {file_name} successfully transferred to frontend")
-            return {"RESULT_CODE": 200, "RESULT_MSG": "File transferred successfully"}
-
-        except requests.exceptions.RequestException as e:
-            logging.error(f"Failed to send file to frontend: {str(e)}")
-            raise HTTPException(status_code=500, detail=f"Request to frontend failed: {str(e)}")
-
-    except Exception as e:
-        logging.error(f"Unexpected error during file transfer: {str(e)}")
-        raise HTTPException(status_code=500, detail=f"Unexpected error: {str(e)}")
 
 def replace_placeholder_in_cell(cell, placeholder, replacement):
     for paragraph in cell.paragraphs:
@@ -132,7 +104,7 @@ def process_meeting_minutes(doc_s_no):
     output_path = f"doc_conv/회의록_{doc_s_no}.docx"
     doc.save(output_path)
 
-    push_to_nextjs(output_path, f"회의록_{doc_s_no}.docx")
+    push.push_to_nextjs(output_path, f"회의록_{doc_s_no}.docx")
 
     return {"RESULT_CODE": 200, "RESULT_MSG": "Done!"}
 
@@ -185,7 +157,7 @@ def process_summary(doc_s_no):
     output_path = f"doc_conv/개요서_{doc_s_no}.docx"
     doc.save(output_path)
 
-    push_to_nextjs(output_path, f"개요서_{doc_s_no}.docx")
+    push.push_to_nextjs(output_path, f"개요서_{doc_s_no}.docx")
 
     return {"RESULT_CODE": 200, "RESULT_MSG": "Done!"}
 
@@ -231,7 +203,7 @@ def process_reqspec(doc_r_no):
     output_path = f"doc_conv/요구사항_{doc_r_no}.docx"
     doc.save(output_path)
 
-    push_to_nextjs(output_path, f"요구사항_{doc_s_no}.docx")
+    push.push_to_nextjs(output_path, f"요구사항_{doc_s_no}.docx")
 
     return {"RESULT_CODE": 200, "RESULT_MSG": "Done!"}
 
@@ -281,7 +253,7 @@ def process_testcase(doc_t_no):
     output_path = f"doc_conv/테스트케이스_{doc_t_no}.docx"
     doc.save(output_path)
 
-    push_to_nextjs(output_path, f"테스트케이스_{doc_s_no}.docx")
+    push.push_to_nextjs(output_path, f"테스트케이스_{doc_s_no}.docx")
 
     return {"RESULT_CODE": 200, "RESULT_MSG": "Done!", "OUTPUT_PATH": output_path}
 
@@ -334,7 +306,7 @@ def process_report(doc_rep_no):
     output_path = f"doc_conv/보고서_{doc_rep_no}.docx"
     doc.save(output_path)
 
-    push_to_nextjs(output_path, f"보고서_{doc_s_no}.docx")
+    push.push_to_nextjs(output_path, f"보고서_{doc_s_no}.docx")
 
     return {"RESULT_CODE": 200, "RESULT_MSG": "Done!", "OUTPUT_PATH": output_path}
